Проверка соединения с базой данных, редирект на страницу информации

+++ OrderApp/Services/OrderService.cs
@@ -1,4 +1,6 @@
-﻿using OrderApp.Interfaces;
+﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
+using OrderApp.Interfaces;
 using OrderApp.Models;
 
 namespace OrderApp.Services;
@@ -12,45 +14,65 @@ public class OrderService
         _orderRepository = orderRepository;
     }
 
-    public async Task AddOrder(Order curOrder)
+    public async Task<Order> AddOrder(Order curOrder)
     {
-        var guidId = Guid.NewGuid().ToString();
-        curOrder.Id = guidId;
-        int? id;
-        if ((await _orderRepository.GetAllOrders()).Count == 0)
-            id = 1;
+        if (_orderRepository.CanConnect())
+        {
+            var guidId = Guid.NewGuid().ToString();
+            curOrder.Id = guidId;
+            int? id;
+            if ((await _orderRepository.GetAllOrders()).Count == 0)
+                id = 1;
+            else
+                id = (await _orderRepository.GetAllOrders()).Last().UserId + 1;
+            curOrder.UserId = id;
+            await _orderRepository.AddOrder(curOrder);
+            return curOrder;
+        }
         else
-            id = (await _orderRepository.GetAllOrders()).Last().UserId + 1;
-        curOrder.UserId = id;
-        await _orderRepository.AddOrder(curOrder);
+        {
+            return curOrder;
+        }
     }
     
     public async Task<List<OrderClientDTO>> GetAllOrders()
     {
-        var orders = await _orderRepository.GetAllOrders();
-        var ordersClientDto = new List<OrderClientDTO>();
-        foreach (var order in orders)
+        if (_orderRepository.CanConnect())
         {
-            ordersClientDto.Add(order.MapToDto());
+            var orders = await _orderRepository.GetAllOrders();
+            var ordersClientDto = new List<OrderClientDTO>();
+            foreach (var order in orders)
+            {
+                ordersClientDto.Add(order.MapToDto());
+            }
+
+            return ordersClientDto.OrderByDescending(o => o.UserId).ToList();
         }
-        return ordersClientDto.OrderByDescending(o=>o.UserId).ToList();
+        else
+        {
+            return new List<OrderClientDTO>();
+        }
     }
 
     public async Task DeleteOrder(string orderId)
     {
-        await _orderRepository.DeleteOrder(orderId);
+        if (_orderRepository.CanConnect())
+        {
+            await _orderRepository.DeleteOrder(orderId);
+        }
     }
 
-    public async Task<OrderClientDTO> GetSingleOrder(string orderId)
-    {
-        var curOrder = await _orderRepository.GetSingleOrder(orderId);
-        return curOrder.MapToDto();
-    }
-    
     public async Task<Order> GetSingleOrder(int orderUserId)
     {
-        var curOrder = (await _orderRepository.GetAllOrders())
-            .FirstOrDefault(o=>o.UserId==orderUserId);
-        return curOrder;
+        if (_orderRepository.CanConnect())
+        {
+            var curOrder = (await _orderRepository.GetAllOrders())
+                .FirstOrDefault(o => o.UserId == orderUserId);
+            return curOrder;
+        }
+        else
+        {
+            return null;
+        }
     }
 }
